Travis

+++ core/utilities/mcmlln-tools.rb
@@ -63,16 +63,23 @@ module Mcmlln
     end
 
     def self.readjson(inputfile)
-      file = File.open(inputfile, "r:utf-8")
-      content = file.read
-      file.close
-      json_hash = JSON.parse(content)
+      if File.exist?(inputfile)
+        file = File.open(inputfile, "r:utf-8")
+        content = file.read
+        file.close
+        json_hash = JSON.parse(content)
+      else
+        json_hash={}
+      end
       json_hash
     end
 
     def self.write_json(hash, json)
-      finaljson = JSON.pretty_generate(hash)
-      File.open(json, 'w+:UTF-8') { |f| f.puts finaljson }
+    	#the 'unless' prevents Travis from erroring on writing json_log file at end of every script
+      unless ARGV.empty?
+        finaljson = JSON.pretty_generate(hash)
+        File.open(json, 'w+:UTF-8') { |f| f.puts finaljson }
+      end
     end
 
     def self.overwriteFile(file, content)
